+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### v0.7.2
 * After the collection view is reordered via dragging: if the collection has a comparator, sort after adding all the models in the visual order.
+* Never "rerender" the collection view unless it has already been rendered.  For example, adding models to the collection should not render the collection view if it was not previously rendered.
+* Don't listen to events from old collection when setOptions( "collection", newCollection ) is called
+
 ### v0.7.1
 
 * Use css classes to keep track of and determine visibility of an item